Refactor and optimize code across multiple modules
- Cleaned up import statements and organized code structure in `sync.rs`, `apply.rs`, `callbacks.rs`, `effects_msg.rs`, `mod.rs`, `state.rs`, `types.rs`, and `layout.rs`.
- Improved performance by replacing `Vec<FileNode>` with `&mut [FileNode]` in `rename_node_in_place` and `sort_nodes` functions.
- Enhanced readability and maintainability by using type aliases and reducing unnecessary code complexity.
- Added dead code warnings for future-proofing in various modules.
- Updated logic for handling document synchronization and version control in `sync.rs`.
- Optimized version vector implementation in `vector/mod.rs` to use `SmallVec` for better memory management.
- Adjusted data types in `models.rs` to use `u32` instead of `usize` for sequence numbers in operations.
- Improved markdown rendering logic in `markdown.rs` for better performance.
- Added new dependencies in `Cargo.toml` for enhanced functionality.

+++ apps/web/src/components/chat/actions.rs
@@ -66,7 +66,7 @@ pub fn make_on_apply(core: CoreState) -> Callback<String> {
             leptos::logging::warn!("No active doc to apply code.");
             return;
         };
-        let pos = getEditorContent().len();
+        let pos = getEditorContent().len() as u32;
         let op = Op::Insert {
             pos,
             content: code.into(),
